# Re-apply "[bert/RoBERTa] Optimize LayerNorm with explicit vectorization using Vec256"