@jianyuh 2019-12-12
@jamesr66a : could you re-stamp this PR? It was reverted due to some hypothesis issues (false positive). The original PR is https://github.com/pytorch/pytorch/pull/29104. Thanks!